You know when you try to delete 50,000 records and it bombs out?
It turns out it's because we have a memory leak in relations. Here
comes the tough choice. Spend anywhere from 4 hours to 4 weeks patching
the memory leak, missing the GDPR deadline, or just say screw it and
halt the script when it's about to bomb?

LOL, kidding, the choice is easy. Just say screw it. We rarely handle
50,000 records with six level deep relations of substantial size. Plus,
WE ARE RUNNING OUT OF TIME (read this in Jack Bauer's voice).