Empty state check must be AFTER error check.
Joomla is doing it wrong in its core code!

If there's a database error in the model $this->items
will contain boolean false instead of an array. This
means you get an InvalidArgumentException from count()
in the empty state check since boolean is not Countable.

As a result you will get a PHP error masking the DB
error, making it very difficult to debug the issue.